+++ apps/cowswap-frontend/src/modules/tradeFormValidation/pure/TradeFormBlankButton/index.tsx
@@ -1,4 +1,4 @@
-import { ReactElement, useEffect, useRef, useState } from 'react'
+import { ReactNode, useEffect, useRef, useState } from 'react'
 
 import { useMediaQuery } from '@cowprotocol/common-hooks'
 import { TEST_IDS } from '@cowprotocol/test-ids'
@@ -21,6 +21,7 @@ const ActionButton = styled.button<{ hasLongText$: boolean }>`
   color: var(${UI.COLOR_BUTTON_TEXT});
   font-size: ${({ hasLongText$ }) => (hasLongText$ ? '16px' : '18px')};
   font-weight: 600;
+
   border-radius: 16px;
   min-height: 58px;
   text-align: center;
@@ -44,7 +45,7 @@ const ActionButton = styled.button<{ hasLongText$: boolean }>`
 `
 
 export interface TradeFormPrimaryButtonProps {
-  children: ReactElement | string
+  children: ReactNode
   disabled?: boolean
   loading?: boolean
   id?: string
@@ -117,12 +118,10 @@ export function TradeFormBlankButton({
       data-click-event={clickEvent}
     >
       {showLoader ? (
-        <>
-          <LongLoadText>
-            <Trans>Confirm with your wallet</Trans>
-          </LongLoadText>{' '}
+        <LongLoadText>
+          <Trans>Confirm with your wallet</Trans>
           <CenteredDots smaller />
-        </>
+        </LongLoadText>
       ) : (
         <>{children}</>
       )}