+++ libs/permit-utils/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [3.7.0](https://github.com/cowprotocol/cowswap/compare/permit-utils-v3.6.0...permit-utils-v3.7.0) (2026-09-09)
+
+### Minor Changes
+
+- feat: configure batching & change permit logic check (#8014)
+
+- feat(solana): fetch quote from Jupiter (#8075)
+
+### Patch Changes
+
+- fix(solana): derive pda address from sdk from program id (#8103)
+
+- Updated dependencies []:
+  - @cowprotocol/hook-dapp-lib@2.2.14
+
 ## [3.6.0](https://github.com/cowprotocol/cowswap/compare/permit-utils-v3.5.0...permit-utils-v3.6.0) (2026-09-03)
 
 ### Minor Changes